Fixed norm and removed some unused function
