+++ src/executor/executor_utils.h
@@ -5,7 +5,6 @@
 # include <executor/redirection.h>
 
 t_exit_code	wait_for_all_processes(int num_of_processes, int *pids);
-void		handle_errors(int error_code, const char *location);
 void		execute_system_command(const t_command *command);
 int			create_new_process(t_multi_pipes *pipes,
 				int current_process, int num_of_process);
