+++ src/parser/command_table.h
@@ -6,7 +6,6 @@
 
 t_command_code	get_command_code(char *input, t_command *command);
 int				get_set_index(const char *start, const char *set);
-t_command		*expand_arg_content(t_command *command, t_buffer *buffer);
 t_bool			is_system_command(const char *input, t_command *command);
 char			*get_set_position(const char *set, char *str_to_check);
 void			cleanup_command_table(t_command *command_table, \
