+++ src/parser/dispatcher.c
@@ -1,15 +1,7 @@
-#include <ctype.h>
-#include <libft.h>
-#include <stdio.h>
 #include <parser/dispatcher.h>
-#include <commands/buffer.h>
 #include <commands/commands.h>
 #include <output/write_to_std.h>
-#include <parser/command_table.h>
-#include <parser/parser.h>
-#include <parser/parse_redirection.h>
 #include <executor/executor_utils.h>
-#include <env/env_utils.h>
 
 t_exit_code	unknown_command(t_command command, t_output_stdout output)
 {
