+++ src/parser/file_utils.c
@@ -1,21 +1,5 @@
-#include <fcntl.h>
 #include <libft.h>
-#include <stdio.h>
-#include <ctype.h>
-#include <executor/executor_utils.h>
 #include <parser/parse_redirection.h>
-#include <readline/readline.h>
-#include <readline/history.h>
-#include <errors/errors.h>
-#include <errno.h>
-
-t_bool	is_valid_filename_char(char c, int redirect_id)
-{
-	if (redirect_id == HERE_DOC)
-		return (FALSE);
-	return (c && !ft_isspace(c)
-		&& !ft_strchr(ALL_TERMINATORS, c));
-}
 
 t_exit_code	open_infile(const char *file, int *in_file)
 {
